+++ packages/account-eip6963/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sophon-labs/account-eip6963
 
+## 2.0.10-next.9
+
+### Patch Changes
+
+- Error checking on react native
+- Updated dependencies
+  - @sophon-labs/account-provider@2.0.10-next.9
+  - @sophon-labs/account-core@2.0.10-next.9
+
 ## 2.0.10-next.8
 
 ### Patch Changes